Move guest account network tests to separate class
