+++ msal/src/test/java/com/microsoft/identity/client/e2e/tests/AcquireTokenAbstractTest.java
@@ -22,11 +22,21 @@
 // THE SOFTWARE.
 package com.microsoft.identity.client.e2e.tests;
 
+import com.microsoft.identity.client.AcquireTokenParameters;
+import com.microsoft.identity.client.AcquireTokenSilentParameters;
+import com.microsoft.identity.client.IAccount;
 import com.microsoft.identity.client.e2e.utils.AcquireTokenTestHelper;
 import com.microsoft.identity.internal.testutils.TestUtils;
 
 import org.junit.After;
 import org.junit.Before;
+
+import java.util.Arrays;
+
+import static com.microsoft.identity.client.e2e.utils.AcquireTokenTestHelper.getAccount;
+import static com.microsoft.identity.client.e2e.utils.AcquireTokenTestHelper.successfulInteractiveCallback;
+import static com.microsoft.identity.client.e2e.utils.AcquireTokenTestHelper.successfulSilentCallback;
+import static com.microsoft.identity.client.e2e.utils.RoboTestUtils.flushScheduler;
 
 public abstract class AcquireTokenAbstractTest extends PublicClientApplicationAbstractTest implements IAcquireTokenTest {
 
@@ -43,5 +53,71 @@ public abstract class AcquireTokenAbstractTest extends PublicClientApplicationAb
         AcquireTokenTestHelper.setAccount(null);
         // remove everything from cache after test ends
         TestUtils.clearCache(SHARED_PREFERENCES_NAME);
+    }
+
+    public void performInteractiveAcquireTokenCall(final String username) {
+        final AcquireTokenParameters parameters = new AcquireTokenParameters.Builder()
+                .startAuthorizationFromActivity(mActivity)
+                .withLoginHint(username)
+                .withScopes(Arrays.asList(mScopes))
+                .withCallback(successfulInteractiveCallback())
+                .build();
+
+
+        mApplication.acquireToken(parameters);
+        flushScheduler();
+    }
+
+    public void performInteractiveAcquireTokenCall(final String username, final String authority) {
+        final AcquireTokenParameters parameters = new AcquireTokenParameters.Builder()
+                .startAuthorizationFromActivity(mActivity)
+                .withLoginHint(username)
+                .withScopes(Arrays.asList(mScopes))
+                .fromAuthority(authority)
+                .withCallback(successfulInteractiveCallback())
+                .build();
+
+
+        mApplication.acquireToken(parameters);
+        flushScheduler();
+    }
+
+    public void performSilentAcquireTokenCall(final String[] scopes) {
+        final AcquireTokenSilentParameters silentParameters = new AcquireTokenSilentParameters.Builder()
+                .forAccount(getAccount())
+                .fromAuthority(getAuthority())
+                .withScopes(Arrays.asList(scopes))
+                .forceRefresh(false)
+                .withCallback(successfulSilentCallback())
+                .build();
+
+        mApplication.acquireTokenSilentAsync(silentParameters);
+        flushScheduler();
+    }
+
+    public void performSilentAcquireTokenCall(final IAccount account) {
+        final AcquireTokenSilentParameters silentParameters = new AcquireTokenSilentParameters.Builder()
+                .forAccount(account)
+                .fromAuthority(account.getAuthority())
+                .withScopes(Arrays.asList(mScopes))
+                .forceRefresh(false)
+                .withCallback(successfulSilentCallback())
+                .build();
+
+        mApplication.acquireTokenSilentAsync(silentParameters);
+        flushScheduler();
+    }
+
+    public void performSilentAcquireTokenCall(final IAccount account, final String authority) {
+        final AcquireTokenSilentParameters silentParameters = new AcquireTokenSilentParameters.Builder()
+                .forAccount(account)
+                .fromAuthority(authority)
+                .withScopes(Arrays.asList(mScopes))
+                .forceRefresh(false)
+                .withCallback(successfulSilentCallback())
+                .build();
+
+        mApplication.acquireTokenSilentAsync(silentParameters);
+        flushScheduler();
     }
 }
